Fix issues with potential visualization
For the 3D surface, plot just the potential instead of the
square of the potential which was what was previously done.
Remove graphical glitches that occur when looking at the surface
plots up close, or when using a large scaling value for
the potential height visualization. Add a constant global shift
factor to the potential if its minimum value falls below zero,
to ensure that the potential always takes positive values when
computing the eigenvectors and eigenvalues.
Include a few more preset potentials, and edit the main HTML file.